backend/requirements.txt
amqp==2.6.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
celery==4.4.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
kombu==4.6.11
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
vine==1.3.0
wcwidth==0.2.6
